1-1-6 v2
Krishna and Jugaad went on call and did the work. Krishna: Worked on the HTML basic structure and coded basic buttons. Jugaad: Worked on the CSS and positions of everything as well as making the app look less basic (even though it is still basic). Have not started on JavaScript functions for the app yet.