Code with image for the dummy.
This new code has some minor modifications, like adding a dummy sprite (FINALLY) and changing the damage flash from a red gradient to a blue gradient. There is also a change to background color. All of the changes are cosmetic. It also outs the image for the sprite and the code into one folder so it is easily executable.